qubo_track/requirements_locked.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements_locked.txt requirements_production.txt
#
annotated-types==0.7.0
    # via pydantic
authlib==1.6.0
    # via
    #   dwave-cloud-client
    #   safety
backports-datetime-fromisoformat==2.0.3
    # via marshmallow
blinker==1.9.0
    # via flask
certifi==2025.6.15
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   dwave-cloud-client
    #   dwave-hybrid
    #   flask
    #   safety
    #   typer
contourpy==1.3.2
    # via matplotlib
cryptography==45.0.5
    # via authlib
cycler==0.12.1
    # via matplotlib
dimod==0.12.20
    # via
    #   -r requirements_production.txt
    #   dwave-hybrid
    #   dwave-inspector
    #   dwave-networkx
    #   dwave-ocean-sdk
    #   dwave-preprocessing
    #   dwave-samplers
    #   dwave-system
    #   dwavebinarycsp
    #   penaltymodel
diskcache==5.6.3
    # via dwave-cloud-client
dparse==0.6.4
    # via
    #   safety
    #   safety-schemas
dwave-cloud-client==0.13.6
    # via
    #   -r requirements_production.txt
    #   dwave-hybrid
    #   dwave-inspector
    #   dwave-ocean-sdk
    #   dwave-system
dwave-gate==0.3.4
    # via
    #   -r requirements_production.txt
    #   dwave-ocean-sdk
dwave-hybrid==0.6.14
    # via
    #   -r requirements_production.txt
    #   dwave-ocean-sdk
dwave-inspector==0.5.4
    # via
    #   -r requirements_production.txt
    #   dwave-ocean-sdk
dwave-networkx==0.8.18
    # via
    #   -r requirements_production.txt
    #   dwave-hybrid
    #   dwave-ocean-sdk
    #   dwave-system
    #   minorminer
dwave-ocean-sdk==8.4.0
    # via -r requirements_production.txt
dwave-optimization==0.6.2
    # via
    #   -r requirements_production.txt
    #   dwave-ocean-sdk
    #   dwave-system
dwave-preprocessing==0.6.9
    # via
    #   -r requirements_production.txt
    #   dwave-hybrid
    #   dwave-ocean-sdk
    #   dwave-system
dwave-samplers==1.6.0
    # via
    #   -r requirements_production.txt
    #   dwave-hybrid
    #   dwave-ocean-sdk
    #   dwave-system
dwave-system==1.32.0
    # via
    #   -r requirements_production.txt
    #   dwave-hybrid
    #   dwave-inspector
    #   dwave-ocean-sdk
dwavebinarycsp==0.3.1
    # via
    #   -r requirements_production.txt
    #   dwave-ocean-sdk
exceptiongroup==1.3.0
    # via pytest
fasteners==0.19
    # via minorminer
filelock==3.12.4
    # via safety
flask==3.1.1
    # via dwave-inspector
fonttools==4.58.4
    # via matplotlib
homebase==1.0.1
    # via
    #   dwave-cloud-client
    #   dwave-system
    #   minorminer
    #   penaltymodel
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via dwave-cloud-client
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   safety
kiwisolver==1.4.8
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
marshmallow==4.0.0
    # via safety
matplotlib==3.9.2
    # via -r requirements_production.txt
mdurl==0.1.2
    # via markdown-it-py
minorminer==0.2.19
    # via
    #   dwave-hybrid
    #   dwave-ocean-sdk
    #   dwave-system
networkx==3.4.2
    # via
    #   -r requirements_production.txt
    #   dwave-hybrid
    #   dwave-networkx
    #   dwave-samplers
    #   dwavebinarycsp
    #   minorminer
    #   penaltymodel
numpy==1.26.4
    # via
    #   -r requirements_production.txt
    #   contourpy
    #   dimod
    #   dwave-gate
    #   dwave-hybrid
    #   dwave-inspector
    #   dwave-networkx
    #   dwave-optimization
    #   dwave-preprocessing
    #   dwave-samplers
    #   dwave-system
    #   matplotlib
    #   minorminer
    #   pandas
    #   patsy
    #   penaltymodel
    #   scipy
    #   statsmodels
orjson==3.10.18
    # via
    #   dwave-cloud-client
    #   dwave-inspector
packaging==25.0
    # via
    #   dparse
    #   dwave-cloud-client
    #   matplotlib
    #   pytest
    #   safety
    #   safety-schemas
    #   statsmodels
pandas==2.0.3
    # via
    #   -r requirements_production.txt
    #   statsmodels
patsy==1.0.1
    # via statsmodels
penaltymodel==1.2.0
    # via
    #   dwave-ocean-sdk
    #   dwavebinarycsp
pillow==11.3.0
    # via matplotlib
plucky==0.4.3
    # via
    #   dwave-cloud-client
    #   dwave-hybrid
pluggy==1.6.0
    # via pytest
psutil==6.0.0
    # via safety
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   dwave-cloud-client
    #   safety
    #   safety-schemas
pydantic-core==2.23.4
    # via pydantic
pygments==2.19.2
    # via rich
pyparsing==3.2.3
    # via matplotlib
pysocks==1.7.1
    # via requests
pytest==8.3.3
    # via -r requirements_production.txt
python-dateutil==2.9.0.post0
    # via
    #   dwave-cloud-client
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
requests[socks]==2.32.4
    # via
    #   dwave-cloud-client
    #   safety
rich==14.0.0
    # via
    #   safety
    #   typer
ruamel-yaml==0.18.14
    # via
    #   safety
    #   safety-schemas
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
safety==3.2.7
    # via -r requirements_production.txt
safety-schemas==0.0.14
    # via safety
scipy==1.15.3
    # via
    #   -r requirements_production.txt
    #   dwave-system
    #   minorminer
    #   penaltymodel
    #   statsmodels
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
statsmodels==0.14.4
    # via -r requirements_production.txt
tomli==2.2.1
    # via
    #   dparse
    #   pytest
typer==0.16.0
    # via safety
typing-extensions==4.14.0
    # via
    #   dwave-cloud-client
    #   exceptiongroup
    #   marshmallow
    #   pydantic
    #   pydantic-core
    #   rich
    #   safety
    #   safety-schemas
    #   typer
    #   typing-inspection
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   dwave-cloud-client
    #   requests
    #   safety
werkzeug==3.1.3
    # via
    #   dwave-cloud-client
    #   flask
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
